requirements.txt
urwid
fuzzyfinder
atomicwrites
pyxdg
Pygments

